requirements.txt
gradio
huggingface_hub
Jinja2
Pillow
torch
transformers
pydantic
pyyaml
